Updating the examples

+++ ArduinoButton/examples/AdvancedDebounceClock/AdvancedDebounceClock.ino
@@ -3,7 +3,7 @@
 // and improve performance if you define ArduinoButtonGlobalClock
 // before including the library
 
-long myGlobalTimeCounter;
+unsigned long myGlobalTimeCounter;
 
 #define ArduinoButtonGlobalClock myGlobalTimeCounter
 #include <ArduinoButton.h>
@@ -11,7 +11,7 @@ long myGlobalTimeCounter;
 Button button1;
 
 // Count how many times button1 has been pressed
-int times;
+unsigned short times;
 
 void setup()
 {
