+++ ArduinoButton/examples/SimpleButtonPressAndRelease/SimpleButtonPressAndRelease.ino
@@ -3,10 +3,10 @@
 Button button1;
 
 // Count how many times button1 has been pressed
-int times;
+unsigned short times;
 
 // Count how many times button1 has been released
-int timesReleased;
+unsigned short timesReleased;
 
 void setup()
 {
